+++ CHANGES.md
@@ -1,3 +1,4 @@
 File for documenting changes/updates from gistemp4.0
 
 - Change in primary data type: passing Pandas dataframes between steps instead of streaming individual records between steps
+- Shifting from the use of an equal area grid to a 2x2 lat x lon grid